Refactor logic for accepting and validating inputs
This commit was made to make the function 'promptUserForCalculatorType'
easier to read by moving the input validation expression into its own
function named 'validateUserInput'.

This commit also moves the creation of a scanner object, and its closing
into the main block so future methods will be able to use the scanner
to accept user input.