Refactor main logic to allow the user to perform another operation
This commit changes the logic in the main block so that when an operation
is complete, the calculator will ask the user if they want to perform
another calculation, and either start from the beginning again or shut
down.

The function 'promptUserToDoAnotherOperation' was created to provide the
prompt to determine if the user wants to 'shut the calculator off' or
not. The function 'validateChoiceToDoAnotherOperation' was also created
to ensure the user inputs their choice correctly.

A bug was also fixed in the function 'validateStandardOperator' where
the multiplication symbol (*) was not properly validated.