Fix bug where user input was falsely flagged as invalid
This commit was made to fix a bug where the scanner in the function
'promptUserToDoAnotherOperation' would flag an input as invalid before
the user would even input anything. The 'nextLine()' method was changed
to the 'next()' method to remedy this, and was also replaced in other
instances where 'nextLine()' was used.

The 'next()' method is the more appropriate choice as all prompts only
ask the user to enter a single word or character. The bug stemmed from
the scanner looking for entire sentences, leading to unanticipated
behavior.