Create function 'validateTrigFunction'
This commit was made to reduce redundant code:
    operator.toLowerCase().equals("sin")
    || operator.toLowerCase().equals("cos")
    || operator.toLowerCase().equals("tan");

This expression appears multiple times. To improve readability it was
placed in a function.

'main' now use this function to determine whether it should take a user
input for the number of operands it will operate on, or
manually set it to 1 as a trig function can only perform its operation
on 1 number.